make map

+++ srcs/mlx.c
@@ -6,7 +6,7 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/03 17:31:39 by minabe            #+#    #+#             */
-/*   Updated: 2023/05/07 11:48:52 by minabe           ###   ########.fr       */
+/*   Updated: 2023/05/07 11:50:54 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ int	deal_key(int key, void *param)
 // map
 void	draw_map(char *map, t_mlx *mlx)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 	size_t	width;
 
 	// (void)map;
